move ChunkCursor to its own file, remove chunk data from cache when chunk reading is finished
